split de la DAO en 1 DAO par type de donnée, split des services suivant le même principe, ajout de deux vues suivant le même principe et suppression des interfaces
