+++ src/DAO/DBProperties.java
@@ -4,6 +4,7 @@ import java.util.Properties;
 import java.io.InputStream;
 import java.io.IOException;
 
+@SuppressWarnings("serial")
 class DBProperties extends Properties{
 
 	private final static String PROPERTIES_FILE_NAME = "db.properties";
