--- src/Service/Page.java
@@ -1,23 +1,0 @@
-package Service;
-
-public class Page {
-	
-	private String type;
-	private int num;
-	
-	public Page() {}
-	
-	public void setType(String type) {
-		this.type = type;
-	}
-	public String getType() {
-		return this.type;
-	}
-	
-	public void setNum(int num) {
-		this.num = num;
-	}
-	public int getNum() {
-		return this.num;
-	}
-}
