--- src/contract/IController.java
@@ -1,8 +1,0 @@
-package contract;
-
-public interface IController {
-
-	boolean action(int commande, boolean boucle);
-
-	void chooseIdDetailcomputer(int commande);
-}
