+++ src/controller/Controller.java
@@ -1,13 +1,14 @@
 package controller;
 
-import contract.IView;
-import contract.IController;
-import contract.IModel;
+import view.View;
+import service.ServiceComputer;
+import service.ServiceCompany;
 
-public class Controller implements IController{
+public class Controller{
 	
-	private IModel model;
-	private IView view;
+	private ServiceComputer serviceComputer;
+	private ServiceCompany serviceCompany;
+	private View view;
 	
 	private final int LIST_COMPUTER = 1;
 	private final int LIST_COMPANY = 2;
@@ -17,8 +18,9 @@ public class Controller implements IController{
 	private final int DELETE_COMPUTER = 6;
 	private final int STOP = 0;
 
-	public Controller(IModel model, IView view) {
-		this.model = model;
+	public Controller(ServiceComputer serviceComputer,ServiceCompany serviceCompany, View view) {
+		this.serviceComputer = serviceComputer;
+		this.serviceCompany = serviceCompany;
 		this.view = view;
 	}
 
@@ -26,12 +28,12 @@ public class Controller implements IController{
 		
 		
 		switch(commande) {
-		case LIST_COMPUTER : this.view.printListComputer(this.model.getListComputer());break;
-		case LIST_COMPANY : this.view.printListCompany(this.model.getListCompany()); break;
-		case DETAIL_COMPUTER : this.view.printAskIdDetailComputer();break;
-		case ADD_COMPUTER : this.view.printAddComputer();break;
-		case UPDATE_COMPUTER: this.view.printUpdateComputer();break;
-		case DELETE_COMPUTER : this.view.printDeleteComputer(); break;
+		case LIST_COMPUTER : this.view.getViewComputer().printListComputer(this.serviceComputer.getListComputer(0));break;
+		case LIST_COMPANY : this.view.getViewCompany().printListCompany(this.serviceCompany.getListCompany(0)); break;
+		case DETAIL_COMPUTER : this.view.getViewComputer().printAskIdDetailComputer();break;
+		case ADD_COMPUTER : this.view.getViewComputer().printAddComputer();break;
+		case UPDATE_COMPUTER: this.view.getViewComputer().printUpdateComputer();break;
+		case DELETE_COMPUTER : this.view.getViewComputer().printDeleteComputer(); break;
 		case STOP : ;
 		default : boucle = false;
 		}
@@ -40,6 +42,14 @@ public class Controller implements IController{
 	}
 
 	public void chooseIdDetailcomputer(int commande) {
-		this.view.printDetailComputer(this.model.getComputer(commande));
+		this.view.getViewComputer().printDetailComputer(this.serviceComputer.getComputer(commande));
+	}
+	
+	public void changePageComputer(int page) {
+		this.view.getViewComputer().printListComputer(this.serviceComputer.getListComputer(page));
+	}
+	
+	public void changePageCompany(int page) {
+		this.view.getViewCompany().printListCompany(this.serviceCompany.getListCompany(page));
 	}
 }
